+++ src/test/java/com/eatthepath/idobfuscator/OffsetIntegerTransformerTest.java
@@ -4,7 +4,7 @@ public class OffsetIntegerTransformerTest extends IntegerTransformerTest {
 
     @Override
     public IntegerTransformer[] getTransformers() {
-        final long[] offsets = new long[] { 1, 17, -876, Long.MAX_VALUE, Long.MIN_VALUE };
+        final int[] offsets = new int[] { 1, 17, -876, Integer.MAX_VALUE, Integer.MIN_VALUE };
         final IntegerTransformer[] transformers = new IntegerTransformer[offsets.length];
 
         for (int i = 0; i < offsets.length; i++) {
